+++ lib/trunk/src/pagespeed/rules/minify_html_test.cc
@@ -127,7 +127,7 @@ class MinifyHtmlTest : public ::testing::Test {
       ASSERT_FALSE(result.has_optimized_content());
     }
 
-    ASSERT_EQ(result.savings().response_bytes_saved(),
+    ASSERT_EQ(static_cast<size_t>(result.savings().response_bytes_saved()),
               strlen(kUnminified) - strlen(kMinified));
     ASSERT_EQ(result.resource_urls_size(), 1);
     ASSERT_EQ(result.resource_urls(0), "http://www.example.com/foo.html");
